+++ src/client.ts
@@ -248,7 +248,7 @@ export class HttpClient {
             return (error as ConnectionError).shouldRetry;
         }
 
-        // Retry on Too-Many-Requests error and internal errors except Service-Unavailable errors
-        return statusCode === 429 || (statusCode >= 500 && statusCode !== 503);
+        // Retry on Too-Many-Requests error and internal errors
+        return statusCode === 429 || statusCode >= 500;
     }
 }